pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
streamlit = "1.28.2"
ipython = "8.17.2"
base256 = "1.0.1"
python-time = "0.3.0"
requests = "2.31.0"
tempfile2 = "0.1.2"
numpy = "1.26.2"
sendgrid = "6.10.0"
moviepy = "1.0.3"
openapi3 = "1.8.2"
openai = "^1.3.3"
python-dotenv = "^1.0.0"
opencv-python-headless = "4.8.1.78"
opencv-contrib-python = "4.8.1.78"
opencv-contrib-python-headless = "4.8.1.78"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"